requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
pdfminer.six==20221105
sentence-transformers==2.2.2
pinecone-client==3.0.0
groq==0.4.1
python-multipart==0.0.6